pyproject.toml
[project]
name = "business-sql-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
